backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
python-chess==1.999
requests==2.31.0
urllib3==2.1.0
python-dotenv==1.0.0
stockfish==3.28.0
celery==5.3.4
redis==5.0.1
email-validator==2.1.0

